moved base objects into a separate JS for cleanness
